+++ src/main/java/su/dreamtime/dreamiki/data/DTCoinsData.java
@@ -8,14 +8,14 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class DTCoinsData
 {
-    private static Map<String, Integer> coinsData = new ConcurrentHashMap<>();
+    private static Map<String, Long> coinsData = new ConcurrentHashMap<>();
 
-    public static synchronized Integer getCoins(String uuid)
+    public static synchronized long getCoins(String uuid)
     {
         return coinsData.get(uuid);
     }
 
-    public static synchronized void rewrite(String uuid, Integer coins)
+    public static synchronized void rewrite(String uuid, long coins)
     {
         coinsData.remove(uuid);
         putNew(uuid, coins);
@@ -26,7 +26,7 @@ public class DTCoinsData
         coinsData.remove(uuid);
     }
 
-    public static synchronized void putNew(String uuid, Integer coins)
+    public static synchronized void putNew(String uuid, long coins)
     {
         if (!coinsData.containsKey(uuid)) {
             coinsData.put(uuid, coins);
